Added create method, to str, and tojson
I Added all three of the methods to both classes and removed some of the looping features because im assuming when we create a new expense or subsriptiont that, that object is singular so there is no reason to prompt the user in a loop. The only down side to this comes in the expense class and its categories and even in subscription. We would have to atleast I think in main account for people adding multiple subscriptions at once which one mean making multiple instances of the class, which I think the logic lies in the main or user manager class. The same is true of keeping track of categories in expenses. If anyone has any ideas about this please implement them